requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
Flask-Migrate>=4.0.5
Flask-JWT-Extended>=4.5.3
Flask-Limiter>=3.5.0
Flask-CORS>=4.0.1
python-dotenv>=1.0.0
pymysql>=1.1.2
passlib[bcrypt]>=1.7.4
pydantic>=2.0
email-validator>=2.0
pytest>=7.0